Обновление команд пользователя.

+++ src/Application/Events.Application.Services/Features/Users/Commands/Delete/DeleteUserHandler.cs
@@ -1,5 +1,7 @@
 ﻿using Events.Application.Services.Shared;
 using Events.Domain.Aggregates.Users;
+using Events.Domain.Aggregates.Users.Errors;
+using Events.Domain.Exceptions;
 using MediatR;
 
 namespace Events.Application.Services.Features.Users.Commands.Delete;
@@ -9,6 +11,9 @@ public sealed class DeleteUserHandler(IRepository<User> userRepository) : IReque
     public async Task Handle(DeleteUserCommand request, CancellationToken cancellationToken)
     {
         var user = await userRepository.GetByIdAsync(request.Id, cancellationToken);
+
+        if (user == null)
+            throw new NotFoundException(UserErrorMessages.UserNotFoundById(request.Id));
 
         await userRepository.DeleteAsync(user, cancellationToken);
     }
